# docs/swagger.yaml
basePath: /
definitions:
  model.BenefitResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      id:
        type: integer
      limitation_type:
        $ref: '#/definitions/model.LimitationTypeResponse'
      name:
        type: string
      plafond:
        type: integer
      plan_type:
        $ref: '#/definitions/model.PlanTypeResponse'
      yearly_max:
        type: integer
    type: object
  model.BenefitResponseWrapper:
    properties:
      access_token:
        type: string
      code:
        type: integer
      data:
        $ref: '#/definitions/model.BenefitResponse'
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PaginationPage'
    type: object
  model.CreateBenefitRequest:
    properties:
      code:
        maxLength: 50
        minLength: 3
        type: string
      detail:
        maxLength: 500
        type: string
      limitation_type_id:
        type: integer
      name:
        maxLength: 255
        minLength: 3
        type: string
      plafond:
        type: number
      plan_type_id:
        type: integer
      yearly_max:
        type: number
    required:
    - code
    - limitation_type_id
    - name
    - plan_type_id
    type: object
  model.ErrorWrapper:
    properties:
      access_token:
        type: string
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PaginationPage'
    type: object
  model.LimitationTypeRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  model.LimitationTypeResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.LimitationTypeResponseWrapper:
    properties:
      access_token:
        type: string
      code:
        type: integer
      data:
        $ref: '#/definitions/model.LimitationTypeResponse'
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PaginationPage'
    type: object
...
  model.PaginationPage:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  model.PlanTypeResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.UpdateBenefitRequest:
    properties:
      code:
        maxLength: 50
        minLength: 3
        type: string
      detail:
        maxLength: 500
        type: string
      id:
        type: integer
      limitation_type_id:
        type: integer
      name:
        maxLength: 255
        minLength: 3
        type: string
      plafond:
        type: number
      plan_type_id:
        type: integer
      yearly_max:
        type: number
    required:
    - code
    - id
    - limitation_type_id
    - name
    - plan_type_id
    type: object
  model.UpdateLimitationTypeRequest:
    properties:
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - id
    - name
    type: object
...
host: 192.168.110.65:3000
info:
  contact: {}
  description: Enter your API Key as X-API-Key header value
  title: Aino Medical API
  version: "1.0"
paths:
...
  /api/v1/benefits:
    get:
      consumes:
      - application/json
      description: Find benefit types by their attributes.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BenefitResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Find benefit types
      tags:
      - Benefit Types
    post:
      consumes:
      - application/json
      description: Create a new benefit type with the provided details.
      parameters:
      - description: Create Benefit Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateBenefitRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BenefitResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Create a new benefit type
      tags:
      - Benefit Types
  /api/v1/benefits/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a benefit type with the provided details.
      parameters:
      - description: Benefit ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BenefitResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Delete a benefit type
      tags:
      - Benefit Types
    get:
      consumes:
      - application/json
      description: Get a benefit type by its ID.
      parameters:
      - description: Benefit ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BenefitResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Get a benefit type by ID
      tags:
      - Benefit Types
    put:
      consumes:
      - application/json
      description: Update a benefit type with the provided details.
      parameters:
      - description: Update Benefit Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateBenefitRequest'
      - description: Benefit ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BenefitResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Update a benefit type
      tags:
      - Benefit Types
  /api/v1/limitation-types:
    get:
      consumes:
      - application/json
      description: Find limitation types by their attributes.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LimitationTypeResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Find limitation types
      tags:
      - Limitation Types
    post:
      consumes:
      - application/json
      description: Create a new limitation type with the provided details.
      parameters:
      - description: Create Limitation Type Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LimitationTypeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LimitationTypeResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Create a new limitation type
      tags:
      - Plan Types
  /api/v1/limitation-types/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a limitation type with the provided details.
      parameters:
      - description: Limitation Type ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LimitationTypeResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Delete a limitation type
      tags:
      - Limitation Types
    get:
      consumes:
      - application/json
      description: Get a limitation type by its ID.
      parameters:
      - description: Limitation Type ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LimitationTypeResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Get a limitation type by ID
      tags:
      - Limitation Types
    put:
      consumes:
      - application/json
      description: Update a limitation type with the provided details.
      parameters:
      - description: Update Limitation Type Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateLimitationTypeRequest'
      - description: Limitation Type ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LimitationTypeResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      security:
      - BearerAuth api_key: []
      summary: Update a limitation type
      tags:
      - Limitation Types
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"